+++ lib/kag/gather/queue.rb
@@ -44,6 +44,18 @@ module KAG
       def reset
         self.players = {}
       end
+
+      def check_for_afk(gather)
+        max_idle = (KAG::Config.instance[:idle][:max] or 1800)
+        max_idle = max_idle.to_i
+        self.players.each do |authname,user|
+          user.refresh
+          if user.idle > max_idle
+            self.remove(user)
+            gather.send_channels_msg "Removed #{user.authname} from queue (#{KAG::Gather::Match.type_as_string}) for being idle too long [#{self.length}]"
+          end
+        end
+      end
     end
   end
 end